update some tests

+++ src/main/java/com/ojins/chatbot/service/LuceneIndexer.java
@@ -8,15 +8,14 @@ import org.apache.lucene.document.Field;
 import org.apache.lucene.document.TextField;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
-import org.apache.lucene.queryparser.classic.QueryParser;
+import org.apache.lucene.index.Term;
 import org.apache.lucene.search.Query;
+import org.apache.lucene.search.TermQuery;
 import org.apache.lucene.store.Directory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.Set;
 
 /**
@@ -40,14 +39,14 @@ public class LuceneIndexer {
     public LuceneIndexer(Directory index, Set<QAState> qaStates) {
         this.index = index;
         try {
-            addManyQAState(qaStates, false);
+            addManyQAState(qaStates, false, false);
         } catch (IOException ex) {
             ex.printStackTrace();
             LOG.error("something wrong when adding QAState");
         }
     }
 
-    private void indexQAState(IndexWriter w, QAState qaState) {
+    private void indexQAState(IndexWriter w, QAState qaState, boolean overwrite) {
         // Each qastate is map to a set of docs. each of them has different question
         // but they all have the same answer
         // consider each doc as a emission path, they all lead to the same state.
@@ -56,9 +55,10 @@ public class LuceneIndexer {
             doc.add(new TextField("Question", p, Field.Store.YES));
             doc.add(new TextField("Answer", qaState.getAnswers().get(0), Field.Store.YES));
             try {
-                Query q = new QueryParser("Question", chineseAnalyzer)
-                        .parse(QueryParser.escape(p));
-                w.deleteDocuments(q);
+                if (overwrite) {
+                    Query q = new TermQuery(new Term("Question", p));
+                    w.deleteDocuments(q);
+                }
                 w.addDocument(doc);
                 w.forceMergeDeletes();
                 w.flush();
@@ -68,7 +68,7 @@ public class LuceneIndexer {
         });
     }
 
-    public void addManyQAState(Set<QAState> qaStates, boolean append) throws IOException {
+    private void addManyQAState(Set<QAState> qaStates, boolean append, boolean overwrite) throws IOException {
         if (qaStates.isEmpty()) return;
         IndexWriter w = new IndexWriter(index, new IndexWriterConfig(chineseAnalyzer));
         if (!append) {
@@ -76,15 +76,22 @@ public class LuceneIndexer {
             w.commit();
         }
         qaStates.forEach(p -> {
-            indexQAState(w, p);
+            indexQAState(w, p, overwrite);
         });
         w.commit();
         w.close();
     }
 
     public boolean addQAState(QAState qaState) {
+        return addQAState(qaState, false);
+    }
+
+    public boolean addQAState(QAState qaState, boolean overwrite) {
         try {
-            addManyQAState(new HashSet<>(Collections.singletonList(qaState)), true);
+            IndexWriter w = new IndexWriter(index, new IndexWriterConfig(chineseAnalyzer));
+            indexQAState(w, qaState, overwrite);
+            w.commit();
+            w.close();
             return true;
         } catch (IOException ex) {
             ex.printStackTrace();
